Split out parsing functions and improve naming
